created BattleManager.cs
This C# example shows the limits of my C# abilities, although it is more of a "for fun" project that I have been working on. The code is very messy and unrefined as a result, but I felt it should be displayed to show not only my ability to utilize advanced functions of C# such as switches and enums, but also that I have a passion for the field and spend my free time refining my skills.